@@ -1,7 +1,7 @@
 {
   "name": "mongoose",
   "description": "Mongoose MongoDB ODM",
-  "version": "8.4.4",
+  "version": "8.4.5",
   "author": "Guillermo Rauch <guillermo@learnboost.com>",
   "keywords": [
     "mongodb",
@@ -28,8 +28,8 @@
     "sift": "17.1.3"
   },
   "devDependencies": {
-    "@babel/core": "7.24.6",
-    "@babel/preset-env": "7.24.6",
+    "@babel/core": "7.24.7",
+    "@babel/preset-env": "7.24.7",
     "@typescript-eslint/eslint-plugin": "^6.2.1",
     "@typescript-eslint/parser": "^6.2.1",
     "acquit": "1.3.0",
@@ -55,19 +55,19 @@
     "markdownlint-cli2": "^0.13.0",
     "marked": "4.3.0",
     "mkdirp": "^3.0.1",
-    "mocha": "10.4.0",
+    "mocha": "10.6.0",
     "moment": "2.x",
-    "mongodb-memory-server": "9.3.0",
+    "mongodb-memory-server": "9.4.0",
     "ncp": "^2.0.0",
     "nyc": "15.1.0",
     "pug": "3.0.3",
     "q": "1.5.1",
     "sinon": "18.0.0",
     "stream-browserify": "3.0.0",
-    "tsd": "0.31.0",
-    "typescript": "5.4.5",
-    "uuid": "9.0.1",
-    "webpack": "5.91.0"
+    "tsd": "0.31.1",
+    "typescript": "5.5.3",
+    "uuid": "10.0.0",
+    "webpack": "5.92.1"
   },
   "directories": {
     "lib": "./lib/mongoose"
