@@ -1,27 +1,42 @@
-require('dotenv').config()
-const express = require('express')
-const mongoose = require('mongoose')
-const cors = require('cors')
-const server = express();
-const router  = require('./routes/auth')
+// packages
+import path from "path";
+import express from "express";
+import dotenv from "dotenv";
+import cookieParser from "cookie-parser";
+import cors from 'cors';
+  
+// Utiles
+import connectDB from "./config/db.js";
+import userRoutes from "./routes/userRoutes.js";
+import categoryRoutes from "./routes/categoryRoutes.js";
+import productRoutes from "./routes/productRoutes.js";
+// import uploadRoutes from "./routes/uploadRoutes.js";
+import orderRoutes from "./routes/orderRoutes.js";
 
-server.use(cors())
-server.use(express.urlencoded())
-server.use(express.json())  
-server.use('',router)         
-// console.log(process.env.JWT_SECRET_KEY)
-try {
-     mongoose.connect(`mongodb+srv://vinaychavan2006:${process.env.DB_PASSWORD}@cluster0.3eacfwr.mongodb.net/ecommerce`, {
-        
-    }).then(console.log(' MongoDB connected '))
-    
-} catch (error) {
-    console.error('index',error.message);
-    process.exit(1);
-}
+dotenv.config();
+const port = process.env.PORT || 5000;      
 
+connectDB();
 
+const app = express();
 
-server.listen(5000,()=>{
-    console.log('server is running on port 5000')
-})
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+app.use(cookieParser());
+app.use(cors());
+  
+app.use("/api/users", userRoutes);
+app.use("/api/category", categoryRoutes);
+app.use("/api/products", productRoutes); 
+// app.use("/api/upload", uploadRoutes);
+app.use("/api/orders", orderRoutes);
+ 
+app.get("/api/config/paypal", (req, res) => {
+  res.send({ clientId: process.env.PAYPAL_CLIENT_ID });
+}); 
+      
+const __dirname = path.resolve();
+app.use("/uploads", express.static(path.join(__dirname + "/uploads")));
+
+app.listen(port, () => console.log(`Server running on port: ${port}`));
+ 
