@@ -1,13 +1,15 @@
 import React from 'react'
-import { Outlet } from 'react-router'
+import { useSelector } from 'react-redux'
+import { Outlet,Navigate } from 'react-router'
 
 const PrivateRoute = () => {
+  const {userInfo} = useSelector(state=>state.auth)
 
   //Auth logic
   console.log('rendered PrivateRoute')
   return (
     <div className='w-full h-full bg-slate-950'>
-      <Outlet/>
+      { userInfo?<Outlet />:<Navigate to="/login"  />}
     </div>
   )
 }
