@@ -1,31 +1,52 @@
 import React, { useState } from "react";
-import { Link, redirect } from "react-router-dom";
+import { Link, redirect, useNavigate } from "react-router-dom";
+import {ToastContainer, toast } from "react-toastify";
+import 'react-toastify/dist/ReactToastify.css';
+import Loader from '../../components/Loader'
 import "./Register.css";
 import axios from 'axios'
+import { useDispatch } from "react-redux";
+import {setCredentials} from '../../redux/features/auth/authSlice'
+import { useRegisterMutation } from "../../redux/api/usersApiSlice";
 const Register = () => {
+  const navigate = useNavigate()
+  const dispatch = useDispatch()
+  const [register, {isLoading, isError, isSuccess, error}] = useRegisterMutation()
   const [formData, setFormData] = useState({
-    name: "",
+    username: "",
     email: "",
     password: "",
+    confirmPassword:""
   });
   const handleChange = (e) => {
     e.preventDefault()
-    console.log('change')
+    // console.log('change')
     setFormData({ ...formData, [e.target.name]: e.target.value });
   };
   const handleSubmit = async (e) => {
-    console.log('sub')
+    console.log('submit clicked')
     e.preventDefault();
+    if(formData.password !== formData.confirmPassword){
+      console.log('pasword',formData.password,'cnfpass',formData.confirmPassword)
+      toast.error('Passwords do not match')
+      return
+    }
     try {
-      const response = await axios.post(
-        "http://localhost:5000/register",
-        formData
-      );
+      const response = await register(formData)
+      if(response.error){
+        console.log(response)
+        toast.error(response.error.data)
+        return
+      }
+      dispatch(setCredentials({...response}))
+      toast.success('Registered Successfully')
+      navigate('/')
       console.log(response);
     } catch (err) {
+      toast.error('Unable to register');
       console.log(err);
     }
-    redirect('/')
+    
   };
   console.log("rendered Register");
   return (
@@ -39,7 +60,7 @@ const Register = () => {
               <br />
               <input
                 type="text"
-                name="name"
+                name="username"
                 onChange={(e)=>handleChange(e)}
                 id="name"
                 placeholder="Enter Name"
@@ -71,6 +92,8 @@ const Register = () => {
               <label htmlFor="cnf-password">Confirm Password</label>
               <br />
               <input
+                onChange={(e)=>handleChange(e)}
+                name="confirmPassword"
                 type="password"
                 id="cnf-password"
                 placeholder="Confirm Password"
@@ -78,9 +101,9 @@ const Register = () => {
             </div>
 
             <button onClick={(e) => handleSubmit(e)} className="signup-button">
-              Register
+            {isLoading ? "Registering..." : "Register"}
             </button>
-
+            {isLoading && <Loader className="mt-4" />}
             <div className="new-user">
               Already Have an Account ?
               <Link className="Link" to="/login">
