@@ -1,6 +1,7 @@
 import React, { useState } from "react";
-import { MdDone } from "react-icons/md";
+
 import { useLocation, useNavigate } from "react-router";
+import ProgressSteps from "../../components/ProgressSteps";
 const Shipping = () => {
   const navigate = useNavigate()
   const {state} = useLocation()
@@ -20,29 +21,7 @@ const Shipping = () => {
     <>
       <div className="w-screen min-h-screen bg-gray-950 flex flex-col items-center">
         <div className="w-4/5 rounded-md flex flex-col items-center p-3 min-h-fit bg-slate-800">
-          <div className="w-1/2 h-fit p-2 pl-16 pr-16 flex justify-between items-center">
-            <div className="flex flex-col items-center">
-              <div className="w-8 h-8 rounded-full bg-green-500 text-white font-bold text-xl box-border flex items-center justify-center">
-                <MdDone />
-              </div>
-              <p className="text-green-500 text-sm">Login</p>
-            </div>
-            <div className="w-[60px] bg-green-400 h-1 mb-5"></div>
-            <div className="flex flex-col items-center">
-              <div className="w-8 h-8 rounded-full bg-green-500 text-white font-bold text-xl p-1 box-border flex items-center justify-center">
-                <MdDone />
-              </div>
-              <p className="text-green-500 text-sm">Shipping</p>
-            </div>
-            <div className="w-[60px] bg-green-400 h-1 mb-5"></div>
-            <div className="flex flex-col items-center">
-              <div className="w-8 h-8 rounded-full bg-white text-gray-400 font-bold text-xl p-1 box-border flex items-center justify-center">
-                3
-              </div>
-              <p className="text-white text-sm">Summary</p>
-            </div>
-            
-          </div>
+          <ProgressSteps step2='true'/>
           <div className="w-[50%] rounded-md box-border p-4 gap-5 h-fit flex flex-col justify-between bg-slate-700 text-white font-semibold">
             <h1 className="text-center">Shipping</h1>
             <label htmlFor="address">
