@@ -1,5 +1,5 @@
 import React from 'react'
-
+// I HAVE ADDED THIS COMPONENT IN NAVIGATION.JSX ITSELF
 const FavoritesCount = () => {
   console.log('rendered FavoritesCpount')
   return (
