@@ -1,6 +1,33 @@
-import React from "react";
+import React, { useEffect, useRef } from "react";
+import { useLocation } from "react-router";
+import { useProfileMutation } from "../../redux/api/usersApiSlice.js";
 
 const Profile = () => {
+  const [profile] = useProfileMutation();
+  const username = useRef();
+  const email = useRef();
+  const password = useRef();
+  const { state } = useLocation();
+  if (state) {
+    console.log(state);
+    // console.log(username.current.value)
+    const setRef = (state) => {
+      username.current.value = state.data.username;
+      email.current.value = state.data.email;
+    };
+    useEffect(() => {
+      setRef(state);
+    }, []);
+  }
+  async function handleProfileUpdate(e) {
+    e.preventDefault();
+    const res = await profile({
+      username: username.current.value,
+      email: username.current.email,
+      password: password.current.value,
+    });
+    console.log(res)
+  }
   return (
     <>
       {/* <div className="w-screen min-h-screen bg-gray-950 ml-[58px] ">
@@ -21,10 +48,10 @@ const Profile = () => {
                 Name
               </label>
               <input
+                ref={username}
                 type="text"
                 placeholder="Enter name"
                 className="w-full px-4 py-2 bg-gray-700 text-white rounded-md focus:outline-none focus:ring-2 focus:ring-pink-600"
-                defaultValue=""
               />
             </div>
             <div className="mb-4">
@@ -33,9 +60,9 @@ const Profile = () => {
               </label>
               <input
                 type="email"
+                ref={email}
                 placeholder="Enter email"
                 className="w-full px-4 py-2 bg-gray-700 text-white rounded-md focus:outline-none focus:ring-2 focus:ring-pink-600"
-                defaultValue=""
               />
             </div>
             <div className="mb-4">
@@ -44,6 +71,7 @@ const Profile = () => {
               </label>
               <input
                 type="password"
+                ref={password}
                 placeholder="Enter password"
                 className="w-full px-4 py-2 bg-gray-700 text-white rounded-md focus:outline-none focus:ring-2 focus:ring-pink-600"
                 defaultValue=""
@@ -62,6 +90,7 @@ const Profile = () => {
             </div>
             <div className="flex justify-between mt-6">
               <button
+                onClick={handleProfileUpdate}
                 type="submit"
                 className="bg-pink-500 text-white py-2 px-6 rounded-md hover:bg-pink-600 transition duration-200 ease-in-out"
               >
