@@ -1,4 +1,4 @@
-export const BASE_URL = "";
+export const BASE_URL = "http://localhost:8080";
 export const USERS_URL = "/api/users";
 export const CATEGORY_URL = "/api/category";
 export const PRODUCT_URL = "/api/products";
